+++ src/js/snowflakes.js
@@ -1,5 +1,5 @@
 import Flake from './flake';
-import {setStyle} from './utils';
+import {setStyle, showElement, hideElement} from './utils';
 
 const
     mainStyle = '{MAIN_STYLE}',
@@ -14,12 +14,16 @@ class Snowflakes {
      * @param {DOMElem} [params.container=document.body]
      * @param {number} [params.count=50]
      * @param {number} [params.color="#5ECDEF"]
-     * @param {number} [params.height=height of container]
+     * @param {number} [params.minOpacity=0.6]
+     * @param {number} [params.maxOpacity=1]
+     * @param {number} [params.minSize=8]
+     * @param {number} [params.maxSize=18]
+     * @param {boolean} [params.rotation=true]
      * @param {number} [params.speed=1]
-     * @param {number} [params.stop=false]
-     * @param {boolean} [useRotate=true]
-     * @param {boolean} [useScale=true]
+     * @param {boolean} [params.stop=false]
+     * @param {number} [params.types=6]
      * @param {number} [params.width=width of container]
+     * @param {number} [params.height=height of container]
      * @param {boolean} [params.wind=true]
      * @param {number} [params.zIndex=9999]
      */
@@ -29,13 +33,13 @@ class Snowflakes {
         this._flakes = [];
         this._isBody = this.params.container === document.body;
 
-        const container = document.createElement('div');
+        const container = this._container = document.createElement('div');
+
         container.classList.add('snowflakes');
         this._isBody && container.classList.add('snowflakes_body');
         setStyle(container, { zIndex: this.params.zIndex });
 
         this.params.container.appendChild(container);
-        this._container = container;
 
         if (this.params.stop) {
             this.stop();
@@ -50,13 +54,16 @@ class Snowflakes {
         this._onResize = () => {
             this._winHeight = this._getWindowHeight();
             const height = this._height();
-            setStyle(this._container, {display: 'none'});
-            this._flakes.forEach(flake => flake.resize(height, this.params.speed));
+            hideElement(container);
+            this._flakes.forEach(flake => flake.resize(height, this.params));
             this._updateAnimationStyle();
-            setStyle(this._container, {display: 'block'});
+            showElement(container);
         };
 
-        this._imagesStyleNode = this._injectStyle(imagesStyle.replace(/%7Bcolor%7D/g, encodeURIComponent(this.params.color)));
+        if (imagesStyle) {
+            this._imagesStyleNode = this._injectStyle(imagesStyle.replace(/%7Bcolor%7D/g, encodeURIComponent(this.params.color)));
+        }
+
         this._animationStyleNode = this._injectStyle(this._getAnimationStyle());
 
         window.addEventListener('resize', this._onResize, false);
@@ -106,12 +113,16 @@ class Snowflakes {
             ['color', '#5ECDEF'],
             ['container', document.body],
             ['count', 50],
-            ['height'],
             ['speed', 1],
             ['stop', false],
-            ['useRotate', true],
-            ['useScale',  true],
+            ['rotation', true],
+            ['minOpacity', 0.6],
+            ['maxOpacity', 1],
+            ['minSize', 8],
+            ['maxSize', 18],
+            ['types', 6],
             ['width'],
+            ['height'],
             ['wind', true],
             ['zIndex', 9999]
         ].forEach(function(item) {
@@ -127,16 +138,13 @@ class Snowflakes {
     }
 
     _getAnimationStyle() {
-        const
-            maxSize = Flake.maxSize,
-            minSize = Flake.minSize,
-            height = (this._height() + maxSize) + 'px';
+        const height = (this._height() + this.params.maxSize) + 'px';
 
         let css = `@-webkit-keyframes snowflake_y{from{-webkit-transform:translateY(0px)}to{-webkit-transform:translateY(${height});}}
 @keyframes snowflake_y{from{transform:translateY(0px)}to{transform:translateY(${height})}}`;
 
-        for (let i = minSize; i <= maxSize; i++) {
-            let left = (i - minSize) * 4 + 'px';
+        for (let i = 0; i <= Flake.maxInnerSize; i++) {
+            let left = (Flake.calcSize(i, this.params) - this.params.minSize) * 4 + 'px';
             css += `@-webkit-keyframes snowflake_x_${i}{from{-webkit-transform:translateX(0px)}to{-webkit-transform:translateX(${left});}}
 @keyframes snowflake_x_${i}{from{transform:translateX(0px)}to{transform:translateX(${left})}}`;
         }
@@ -188,7 +196,7 @@ class Snowflakes {
 
     _height() {
         return this.params.height ||
-            (this._isBody ? this._winHeight : this.params.container.offsetHeight + Flake.maxSize);
+            (this._isBody ? this._winHeight : this.params.container.offsetHeight + this.params.maxSize);
     }
 
     _getWindowHeight() {